AliExpress-Affiliate-Bot/requirements.txt
# Telegram bot
python-telegram-bot[ext]>=21.0

# HTTP (async)
httpx>=0.27.0

# LLM clients
groq>=0.5.0
anthropic>=0.30.0

# Environment
python-dotenv>=1.0.0

# Database (stdlib sqlite3 is used directly — no ORM)

# Utilities
aiosqlite>=0.20.0
